refactor(role): consolidate Role definitions into user.py files as per Arino's suggestion
- Move Role model from models/role.py to models/user.py
- Move Role schemas from schemas/role.py to schemas/user.py
- Update all imports across repository, service, and API layers
- Remove separate role.py files to follow project conventions
- Maintain all functionality with improved file organization

This follows the feedback from PR #94 to keep Role definitions
co-located with User definitions for better cohesion and
consistency with existing project structure.
